requirements.txt
discord
python-dotenv
pylint
urllib3